delete unused Marsh files and move MarshOrUnMarsh class to the JAXB folder
