+++ src/main/java/com/solvd/sql/jaxb/MarshOrUnMarsh.java
@@ -1,4 +1,4 @@
-package com.solvd.util;
+package com.solvd.sql.jaxb;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
